+++ src/main/kotlin/service/store/data/Model.kt
@@ -3,6 +3,6 @@ package service.store.data
 data class ThinPokemon(
     val id: Int,
     val name: String,
-    var xp: Int,
+    var xp: Long,
     val sprite: String?
 )
